refactor test case
