+++ src/Microsoft.Docs.Build.Analyzer.Test/Verifiers/CSharpAnalyzerVerifier.cs
@@ -1,4 +1,4 @@
-﻿using System.Threading;
+using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp.Testing;
@@ -8,11 +8,12 @@ using Microsoft.CodeAnalysis.Testing.Verifiers;
 
 namespace Microsoft.Docs.Build.Analyzers.Test
 {
-    public static partial class CSharpAnalyzerVerifier<TAnalyzer>
+    public static class CSharpAnalyzerVerifier<TAnalyzer>
         where TAnalyzer : DiagnosticAnalyzer, new()
     {
-        /// <inheritdoc cref="AnalyzerVerifier{TAnalyzer, TTest, TVerifier}.Diagnostic()"/>
+#pragma warning disable CA1000 // Do not declare static members on generic types
         public static DiagnosticResult Diagnostic()
+
             => CSharpAnalyzerVerifier<TAnalyzer, MSTestVerifier>.Diagnostic();
 
         /// <inheritdoc cref="AnalyzerVerifier{TAnalyzer, TTest, TVerifier}.Diagnostic(string)"/>
@@ -33,6 +34,29 @@ namespace Microsoft.Docs.Build.Analyzers.Test
 
             test.ExpectedDiagnostics.AddRange(expected);
             await test.RunAsync(CancellationToken.None);
+        }
+#pragma warning restore CA1000 // Do not declare static members on generic types
+
+        public class Test : CSharpAnalyzerTest<TAnalyzer, MSTestVerifier>
+        {
+            public Test()
+            {
+                SolutionTransforms.Add((solution, projectId) =>
+                {
+                    var project = solution.GetProject(projectId);
+                    if (project != null)
+                    {
+                        var compilationOptions = project.CompilationOptions;
+                        if (compilationOptions != null)
+                        {
+                            compilationOptions = compilationOptions.WithSpecificDiagnosticOptions(
+                            compilationOptions.SpecificDiagnosticOptions.SetItems(CSharpVerifierHelper.NullableWarnings));
+                            solution = solution.WithProjectCompilationOptions(projectId, compilationOptions);
+                        }
+                    }
+                    return solution;
+                });
+            }
         }
     }
 }
