--- src/Microsoft.Docs.Build.Analyzer.Test/Verifiers/CSharpAnalyzerVerifier`1+Test.cs
@@ -1,26 +1,0 @@
-﻿using Microsoft.CodeAnalysis.CSharp.Testing;
-using Microsoft.CodeAnalysis.Diagnostics;
-using Microsoft.CodeAnalysis.Testing.Verifiers;
-
-namespace Microsoft.Docs.Build.Analyzers.Test
-{
-    public static partial class CSharpAnalyzerVerifier<TAnalyzer>
-        where TAnalyzer : DiagnosticAnalyzer, new()
-    {
-        public class Test : CSharpAnalyzerTest<TAnalyzer, MSTestVerifier>
-        {
-            public Test()
-            {
-                SolutionTransforms.Add((solution, projectId) =>
-                {
-                    var compilationOptions = solution.GetProject(projectId).CompilationOptions;
-                    compilationOptions = compilationOptions.WithSpecificDiagnosticOptions(
-                        compilationOptions.SpecificDiagnosticOptions.SetItems(CSharpVerifierHelper.NullableWarnings));
-                    solution = solution.WithProjectCompilationOptions(projectId, compilationOptions);
-
-                    return solution;
-                });
-            }
-        }
-    }
-}
